Support multiple storage clients for multi-clouds
