+++ go/base/blobstore/azure/BUILD.bazel
@@ -1,0 +1,30 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "client.go",
+        "config.go",
+        "module.go",
+    ],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/base/blobstore/azure",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//go/base/blobstore:go_default_library",
+        "@org_uber_go_config//:go_default_library",
+        "@org_uber_go_fx//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "client_test.go",
+        "config_test.go",
+        "module_test.go",
+    ],
+    embed = [":go_default_library"],
+    deps = [
+        "@org_uber_go_config//:go_default_library",
+    ],
+)
