+++ go/base/blobstore/minio/module_test.go
@@ -41,10 +41,12 @@ func TestNewClient_MultipleProviders_CreatesMultipleClients(t *testing.T) {
 				AwsSecretAccessKey: "testsecret",
 				AwsEndpointUrl:     "s3.amazonaws.com",
 			},
-			"azure-dev": {
-				Type:                "azure",
-				AzureStorageAccount: "testaccount",
-				AzureSASToken:       "testtoken",
+			"aws-dev": {
+				Type:               "s3",
+				AwsRegion:          "us-east-1",
+				AwsAccessKeyId:     "devkey",
+				AwsSecretAccessKey: "devsecret",
+				AwsEndpointUrl:     "localhost:9000",
 			},
 		},
 	}
@@ -75,16 +77,16 @@ func TestNewClient_MultipleProviders_CreatesMultipleClients(t *testing.T) {
 	if !providerKeys["aws-prod"] {
 		t.Error("expected aws-prod provider key")
 	}
-	if !providerKeys["azure-dev"] {
-		t.Error("expected azure-dev provider key")
+	if !providerKeys["aws-dev"] {
+		t.Error("expected aws-dev provider key")
 	}
 
-	// Should have both schemes
+	// Should have s3 scheme
 	if !schemes["s3"] {
 		t.Error("expected s3 scheme")
 	}
-	if !schemes["abfss"] {
-		t.Error("expected abfss scheme")
+	if len(schemes) != 1 {
+		t.Errorf("expected only 1 scheme, got %d: %v", len(schemes), schemes)
 	}
 }
 
@@ -132,84 +134,32 @@ func TestNewS3ClientWithKey_StaticCredentials(t *testing.T) {
 	}
 }
 
-func TestNewAzureClientWithKey_Success(t *testing.T) {
-	config := StorageProvider{
-		Type:                "azure",
-		AzureStorageAccount: "testaccount",
-		AzureSASToken:       "testtoken",
-	}
-
-	client, err := newAzureClientWithKey("test-azure", config)
-	if err != nil {
-		t.Fatalf("expected no error, got %v", err)
-	}
-
-	blobClient := client.BlobStoreClient
-	if blobClient.Scheme() != "abfss" {
-		t.Errorf("expected scheme 'abfss', got %q", blobClient.Scheme())
-	}
-
-	if providerClient, ok := blobClient.(interface{ ProviderKey() string }); ok {
-		if providerClient.ProviderKey() != "test-azure" {
-			t.Errorf("expected provider key 'test-azure', got %q", providerClient.ProviderKey())
-		}
-	} else {
-		t.Error("expected client to implement ProviderKey method")
-	}
-}
-
-func TestNewAzureClientWithKey_MissingStorageAccount(t *testing.T) {
-	config := StorageProvider{
-		Type:          "azure",
-		AzureSASToken: "testtoken",
-		// Missing AzureStorageAccount
-	}
-
-	_, err := newAzureClientWithKey("test-azure", config)
-	if err == nil {
-		t.Fatal("expected error for missing storage account")
-	}
-
-	expectedError := "azure storage account is required for provider test-azure"
-	if err.Error() != expectedError {
-		t.Errorf("expected error %q, got %q", expectedError, err.Error())
-	}
-}
-
-func TestNewAzureClientWithKey_MissingSASToken(t *testing.T) {
-	config := StorageProvider{
-		Type:                "azure",
-		AzureStorageAccount: "testaccount",
-		// Missing AzureSASToken
-	}
-
-	_, err := newAzureClientWithKey("test-azure", config)
-	if err == nil {
-		t.Fatal("expected error for missing SAS token")
-	}
-
-	expectedError := "azure SAS token is required for provider test-azure"
-	if err.Error() != expectedError {
-		t.Errorf("expected error %q, got %q", expectedError, err.Error())
-	}
-}
-
-func TestNewClient_UnsupportedProviderType(t *testing.T) {
+func TestNewClient_NonS3Providers_SkipsNonS3(t *testing.T) {
 	config := Config{
 		StorageProviders: map[string]StorageProvider{
-			"gcp-prod": {
-				Type: "gcp", // Unsupported type
+			"aws-dev": {
+				Type:               "s3",
+				AwsAccessKeyId:     "devkey",
+				AwsSecretAccessKey: "devsecret",
+			},
+			"azure-prod": {
+				Type: "azure", // Non-S3 provider should be skipped
 			},
 		},
 	}
 
-	_, err := newClient(config)
-	if err == nil {
-		t.Fatal("expected error for unsupported provider type")
+	clients, err := newClient(config)
+	if err != nil {
+		t.Fatalf("expected no error, got %v", err)
 	}
 
-	expectedError := "unsupported storage provider type: gcp for provider gcp-prod"
-	if err.Error() != expectedError {
-		t.Errorf("expected error %q, got %q", expectedError, err.Error())
+	// Should only create 1 S3 client, skipping the Azure provider
+	if len(clients) != 1 {
+		t.Fatalf("expected 1 S3 client, got %d", len(clients))
+	}
+
+	client := clients[0].BlobStoreClient
+	if client.Scheme() != "s3" {
+		t.Errorf("expected scheme 's3', got %q", client.Scheme())
 	}
 }
